job: move resolutions to tasks utility method to job module
The goal is to make the job and the resolver more closely aligned,
and remove custom code from the 'nrun' plugin, which is temporary
and experimental only and will be removed as soon as possible.
