+++ avocado/core/job.py
@@ -34,7 +34,9 @@ from . import exit_codes
 from . import job_id
 from . import jobdata
 from . import loader
+from . import nrunner
 from . import output
+from . import resolver
 from . import result
 from . import tags
 from . import test
@@ -49,9 +51,61 @@ from .output import LOG_JOB
 from .output import LOG_UI
 from .output import STD_OUTPUT
 from .settings import settings
+from .tags import filter_test_tags_runnable
 
 
 _NEW_ISSUE_LINK = 'https://github.com/avocado-framework/avocado/issues/new'
+
+
+def resolutions_to_tasks(resolutions, config):
+    """
+    Transforms resolver resolutions into tasks
+
+    A resolver resolution
+    (:class:`avocado.core.resolver.ReferenceResolution`) contains
+    information about the resolution process (if it was successful
+    or not) and in case of sucessful resolutions a list of
+    resolutions.  It's expected that the resolution are
+    :class:`avocado.core.nrunner.Runnable`.
+
+    This method transforms those runnables into Tasks
+    (:class:`avocado.core.nrunner.Task`), which will include an
+    unique sequential identification and a status reporting URI.
+    It also performs tag based filtering on the runnables for
+    possibly excluding some of the Runnables.
+
+    :param resolutions: possible multiple resolutions for multiple
+                        references
+    :type resolutions: list of :class:`avocado.core.resolver.ReferenceResolution`
+    :param config: job configuration
+    :type config: dict
+    :returns: the resolutions converted to tasks
+    :rtype: list of :class:`avocado.core.nrunner.Task`
+    """
+
+    tasks = []
+    index = 0
+    resolutions = [res for res in resolutions if
+                   res.result == resolver.ReferenceResolutionResult.SUCCESS]
+    no_digits = len(str(len(resolutions)))
+    for resolution in resolutions:
+        name = resolution.reference
+        for runnable in resolution.resolutions:
+            filter_by_tags = config.get('filter_by_tags')
+            if filter_by_tags:
+                if not filter_test_tags_runnable(
+                        runnable,
+                        filter_by_tags,
+                        config.get('filter_by_tags_include_empty'),
+                        config.get('filter_by_tags_include_empty_key')):
+                    continue
+            if runnable.uri:
+                name = runnable.uri
+            identifier = str(test.TestID(index + 1, name, None, no_digits))
+            tasks.append(nrunner.Task(identifier, runnable,
+                                      [config.get('status_server')]))
+            index += 1
+    return tasks
 
 
 class Job:
